% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{enable}
\alias{enable}
\alias{disable}
\title{Enable/Disable Copr Package Manager}
\usage{
enable()

disable()
}
\description{
Functions to enable or disable the integration of \code{\link{install_copr}}
into \code{\link{install.packages}}. When enabled, packages are installed
transparently from Copr if available, and from CRAN if not.
}
