% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{install_copr}
\alias{install_copr}
\alias{remove_copr}
\title{Install Binary Packages from Copr}
\usage{
install_copr(pkgs)

remove_copr(pkgs)
}
\arguments{
\item{pkgs}{character vector of CRAN names of packages.}
}
\value{
Invisibly, a character vector of the names of packages not available.
}
\description{
Talk to the accompanying D-Bus service to download and install or remove
 packages from Copr repositories.
}
